Refactor how game excutes movement
I redid the movement after lecture because this was the direction I was wanting to go
but did not know how to get currentRoom to be in player. So I have now changed
my code a bit to implement it in a cleaner way. I did save my before refactored
code below, so you can still see my train of thought.